+++ src/mol-plugin-ui/plugin.tsx
@@ -10,7 +10,7 @@ import * as React from 'react';
 import { formatTime } from '../mol-util';
 import { LogEntry } from '../mol-util/log-entry';
 import { PluginReactContext, PluginUIComponent } from './base';
-import { AnimationViewportControls, DefaultStructureTools, LociLabels, StateSnapshotViewportControls, TrajectoryViewportControls, SelectionViewportControls } from './controls';
+import { AnimationViewportControls, DefaultStructureTools, LociLabels, StateSnapshotViewportControls, TrajectoryViewportControls, SelectionViewportControls, ViewportSnapshotDescription } from './controls';
 import { LeftPanelControls } from './left-panel';
 import { SequenceView } from './sequence';
 import { BackgroundTaskProgress, OverlayTaskProgress } from './task';
@@ -168,7 +168,26 @@ class Layout extends PluginUIComponent {
     };
 
     private showDragOverlay = new BehaviorSubject(false);
-    onDragEnter = () => this.showDragOverlay.next(true);
+    onDragEnter = (ev: React.DragEvent<HTMLDivElement>) => {
+        let hasFile = false;
+        if (ev.dataTransfer.items) {
+            for (let i = 0; i < ev.dataTransfer.items.length; i++) {
+                if (ev.dataTransfer.items[i].kind !== 'file') continue;
+                hasFile = true;
+                break;
+            }
+        } else {
+            for (let i = 0; i < ev.dataTransfer.files.length; i++) {
+                if (!ev.dataTransfer.files[i]) continue;
+                hasFile = true;
+                break;
+            }
+        }
+
+        if (hasFile) {
+            this.showDragOverlay.next(true);
+        }
+    };
 
     render() {
         const layout = this.plugin.layout.state;
@@ -254,6 +273,7 @@ export class DefaultViewport extends PluginUIComponent {
                 <AnimationViewportControls />
                 <TrajectoryViewportControls />
                 <StateSnapshotViewportControls />
+                <ViewportSnapshotDescription />
             </div>
             <SelectionViewportControls />
             <VPControls />